+++ keramics-formats/src/hfs/catalog_file.rs
@@ -154,7 +154,7 @@ impl HfsCatalogFile {
             }
             if !is_branch {
                 if key.parent_identifier == parent_identifier {
-                    let mut data_offset: usize = key.size as usize;
+                    let mut data_offset: usize = key.size;
 
                     if self.btree_file.format == HfsFormat::Hfs {
                         let alignment_padding: usize = calculate_alignment_padding(data_offset, 2);
@@ -208,7 +208,7 @@ impl HfsCatalogFile {
                 }
             } else if record_index > 0 {
                 if key.parent_identifier >= parent_identifier {
-                    let data_offset: usize = last_key.size as usize;
+                    let data_offset: usize = last_key.size;
 
                     if data_offset + 4 > last_record_data.len() {
                         return Err(keramics_core::error_trace_new!(format!(
@@ -259,7 +259,7 @@ impl HfsCatalogFile {
                     "Invalid record index value out of bounds"
                 ));
             }
-            let data_offset: usize = last_key.size as usize;
+            let data_offset: usize = last_key.size;
 
             if data_offset + 4 > last_record_data.len() {
                 return Err(keramics_core::error_trace_new!(format!(
@@ -493,7 +493,7 @@ impl HfsCatalogFile {
                 } else if key.size == 0 {
                     false
                 } else {
-                    let mut data_offset: usize = key.size as usize;
+                    let mut data_offset: usize = key.size;
 
                     if self.btree_file.format == HfsFormat::Hfs {
                         let alignment_padding: usize = calculate_alignment_padding(data_offset, 2);
@@ -591,7 +591,7 @@ impl HfsCatalogFile {
                     "Invalid record index value out of bounds"
                 ));
             }
-            let data_offset: usize = last_key.size as usize;
+            let data_offset: usize = last_key.size;
 
             if data_offset + 4 > last_record_data.len() {
                 return Err(keramics_core::error_trace_new!(format!(
@@ -758,7 +758,7 @@ impl HfsCatalogFile {
                     "Invalid record index value out of bounds"
                 ));
             }
-            let data_offset: usize = last_key.size as usize;
+            let data_offset: usize = last_key.size;
 
             if data_offset + 4 > last_record_data.len() {
                 return Err(keramics_core::error_trace_new!(format!(
@@ -825,7 +825,7 @@ impl HfsCatalogFile {
         key: &HfsCatalogKey,
         record_data: &[u8],
     ) -> Result<HfsDirectoryEntry, ErrorTrace> {
-        let mut data_offset: usize = key.size as usize;
+        let mut data_offset: usize = key.size;
 
         if self.btree_file.format == HfsFormat::Hfs {
             let alignment_padding: usize = calculate_alignment_padding(data_offset, 2);
@@ -908,7 +908,7 @@ impl HfsCatalogFile {
         key: &HfsCatalogKey,
         record_data: &[u8],
     ) -> Result<HfsCatalogThreadRecord, ErrorTrace> {
-        let mut data_offset: usize = key.size as usize;
+        let mut data_offset: usize = key.size;
 
         if self.btree_file.format == HfsFormat::Hfs {
             let alignment_padding: usize = calculate_alignment_padding(data_offset, 2);
