+++ keramics-formats/src/ntfs/index.rs
@@ -165,7 +165,7 @@ impl NtfsIndex {
                         ));
                     }
                 };
-                virtual_cluster_number += data_run.number_of_blocks as u64;
+                virtual_cluster_number += data_run.number_of_blocks;
                 virtual_cluster_offset += range_size;
             }
             if cluster_group.last_vcn != 0xffffffffffffffff
